requirements.txt
# =============================================================================
# Insurance RAG System - Python Dependencies
# =============================================================================

# --- Web Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
sse-starlette==2.1.0

# --- Database ---
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
psycopg[binary]==3.2.3
alembic==1.13.0

# --- Task Queue ---
celery[redis]==5.4.0
redis==5.1.0

# --- AI / ML ---
openai==1.50.0
anthropic==0.34.0
pinecone-client==5.0.0
tiktoken==0.8.0

# --- PDF Processing ---
PyMuPDF==1.24.0
pytesseract==0.3.13
Pillow==10.4.0

# --- AWS ---
boto3==1.35.0

# --- Auth ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0

# --- Utilities ---
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
tenacity==9.0.0
structlog==24.4.0

# --- Testing ---
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0

# --- Dev ---
ruff==0.6.0
